+++ recipes/duckx/all/test_package/test_package.cpp
@@ -1,0 +1,8 @@
+#include <iostream>
+#include "duckx.hpp"
+
+int main() {
+    duckx::Document doc{};
+
+    return 0;
+}