+++ recipes/xz_utils/cmake/test_package/test_package.c
@@ -1,0 +1,8 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include <lzma.h>
+
+int main() {
+    printf("LZMA version %s\n", lzma_version_string());
+    return EXIT_SUCCESS;
+}